+++ src/platform_impl/linux/wayland/window/shim.rs
@@ -1,18 +1,25 @@
-use std::cell::Cell;
+use std::cell::{Cell, RefCell};
 use std::sync::{Arc, Mutex};
 
 use sctk::reexports::client::protocol::wl_output::WlOutput;
+use sctk::reexports::client::Attached;
+use sctk::reexports::protocols::staging::xdg_activation::v1::client::xdg_activation_token_v1;
+use sctk::reexports::protocols::staging::xdg_activation::v1::client::xdg_activation_token_v1::XdgActivationTokenV1;
+use sctk::reexports::protocols::staging::xdg_activation::v1::client::xdg_activation_v1::XdgActivationV1;
 
+use sctk::environment::Environment;
 use sctk::window::{Decorations, FallbackFrame, Window};
 
 use crate::dpi::{LogicalPosition, LogicalSize};
 
 use crate::event::WindowEvent;
+use crate::platform_impl::wayland;
+use crate::platform_impl::wayland::env::WinitEnv;
 use crate::platform_impl::wayland::event_loop::WinitState;
 use crate::platform_impl::wayland::seat::pointer::WinitPointer;
 use crate::platform_impl::wayland::seat::text_input::TextInputHandler;
 use crate::platform_impl::wayland::WindowId;
-use crate::window::CursorIcon;
+use crate::window::{CursorIcon, UserAttentionType};
 
 /// A request to SCTK window from Winit window.
 #[derive(Debug, Clone)]
@@ -63,6 +70,11 @@ pub enum WindowRequest {
 
     /// Set IME window position.
     IMEPosition(LogicalPosition<u32>),
+
+    /// Request Attention.
+    ///
+    /// `None` unsets the attention request.
+    Attention(Option<UserAttentionType>),
 
     /// Redraw was requested.
     Redraw,
@@ -150,14 +162,23 @@ pub struct WindowHandle {
 
     /// Text inputs on the current surface.
     text_inputs: Vec<TextInputHandler>,
+
+    /// XdgActivation object.
+    xdg_activation: Option<Attached<XdgActivationV1>>,
+
+    /// Actication token.
+    xdg_activation_token: RefCell<Option<XdgActivationTokenV1>>,
 }
 
 impl WindowHandle {
     pub fn new(
+        env: &Environment<WinitEnv>,
         window: Window<FallbackFrame>,
         size: Arc<Mutex<LogicalSize<u32>>>,
         pending_window_requests: Arc<Mutex<Vec<WindowRequest>>>,
     ) -> Self {
+        let xdg_activation = env.get_global::<XdgActivationV1>();
+
         Self {
             window,
             size,
@@ -167,6 +188,8 @@ impl WindowHandle {
             cursor_visible: Cell::new(true),
             pointers: Vec::new(),
             text_inputs: Vec::new(),
+            xdg_activation,
+            xdg_activation_token: RefCell::new(None),
         }
     }
 
@@ -184,6 +207,48 @@ impl WindowHandle {
                 pointer.confine(surface);
             } else {
                 pointer.unconfine();
+            }
+        }
+    }
+
+    pub fn set_user_attention(&self, request_type: Option<UserAttentionType>) {
+        let xdg_activation = match self.xdg_activation.as_ref() {
+            None => return,
+            Some(xdg_activation) => xdg_activation,
+        };
+
+        match request_type {
+            None => {
+                if let Some(token) = self.xdg_activation_token.take() {
+                    token.destroy();
+                }
+            }
+            Some(_) => {
+                // The token is already present, nothing to set.
+                if self.xdg_activation_token.borrow().is_some() {
+                    return;
+                }
+
+                let xdg_activation_token = xdg_activation.get_activation_token();
+                let surface = self.window.surface();
+                let window_id = wayland::make_wid(surface);
+                let xdg_activation = xdg_activation.clone();
+                xdg_activation_token.quick_assign(move |xdg_token, event, mut dispatch_data| {
+                    let winit_state = dispatch_data.get::<WinitState>().unwrap();
+                    if let Some(window_handle) = winit_state.window_map.get_mut(&window_id) {
+                        let surface = window_handle.window.surface();
+                        if let xdg_activation_token_v1::Event::Done { token } = event {
+                            xdg_activation.activate(token, surface);
+                            *window_handle.xdg_activation_token.borrow_mut() = None;
+                            xdg_token.destroy();
+                        }
+                    }
+                });
+
+                xdg_activation_token.set_surface(surface);
+                xdg_activation_token.commit();
+                self.xdg_activation_token
+                    .replace(Some(xdg_activation_token.detach()));
             }
         }
     }
@@ -360,6 +425,9 @@ pub fn handle_window_requests(winit_state: &mut WinitState) {
                     // We should refresh the frame after resize.
                     let window_update = window_updates.get_mut(window_id).unwrap();
                     window_update.refresh_frame = true;
+                }
+                WindowRequest::Attention(request_type) => {
+                    window_handle.set_user_attention(request_type);
                 }
                 WindowRequest::Redraw => {
                     let window_update = window_updates.get_mut(window_id).unwrap();